Move survey cloning into its own namespace
