+++ indico/modules/events/__init__.py
@@ -234,7 +234,6 @@ def _get_cloners(sender, **kwargs):
     yield clone.EventPersonCloner
     yield clone.EventPersonLinkCloner
     yield clone.EventProtectionCloner
-    yield clone.EventSurveyCloner
 
 
 @signals.event.cloned.connect
